requirements.txt
Django==1.5.1
Markdown==2.3.1
PIL==1.1.7
South==0.7.6
cGPolyEncode==0.1.1
django-profiles==0.2
django-registration==0.8
django-tastypie==0.9.14
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
wsgiref==0.1.2
six==1.3.0
sorl-thumbnail==11.12
